wgc: move authserver and httpclient outside of wgc_api
