+++ wgc/wgc_api.py
@@ -1,7 +1,6 @@
 # (c) 2019-2020 Mikhail Paulyshka
 # SPDX-License-Identifier: MIT
 
-import asyncio
 from collections import namedtuple
 import json
 import logging
@@ -16,110 +15,15 @@ import threading
 from typing import Any, Dict, List
 from urllib.parse import parse_qs
 
-
-import aiohttp
-import aiohttp.web
-import certifi
+import asyncio
 
 from .wgc_application_owned import WGCOwnedApplication
+from .wgc_authserver import WGCAuthorizationServer
 from .wgc_constants import WGCAuthorizationResult, WGCRealms, GAMES_F2P
 from .wgc_keccak import Keccak512
-
-class WGCAuthorizationServer():
-
-    def __init__(self, backend):
-        self.__backend = backend
-        self.__app = aiohttp.web.Application()
-
-        self.__runner = None
-        self.__site = None
-
-        self.__app.add_routes([
-            aiohttp.web.get ('/login'                , self.handle_login_get                    ),
-            aiohttp.web.get ('/login_failed'         , self.handle_login_failed_get             ),
-            aiohttp.web.get ('/2fa'                  , self.handle_2fa_get                      ),
-            aiohttp.web.get ('/2fa_failed'           , self.handle_2fa_failed_get               ),
-            aiohttp.web.get ('/finished'             , self.handle_finished_get                 ),
-            aiohttp.web.get ('/unsupported_platform' , self.handle_unsupported_platform_get     ),
-            
-            aiohttp.web.post('/login'   , self.handle_login_post  ),   
-            aiohttp.web.post('/2fa'     , self.handle_2fa_post    ),
-        ])
-    
-    async def start(self, host, port):
-        self.__runner = aiohttp.web.AppRunner(self.__app)
-        await self.__runner.setup()
-    
-        self.__site = aiohttp.web.TCPSite(self.__runner, host, port)
-        await self.__site.start()    
-
-    async def shutdown(self):    
-        await self.__runner.cleanup()
-
-    async def handle_login_get(self, request):
-        return aiohttp.web.FileResponse(os.path.join(os.path.dirname(os.path.realpath(__file__)),'html/login.html'))
-
-    async def handle_unsupported_platform_get(self, request):
-        return aiohttp.web.FileResponse(os.path.join(os.path.dirname(os.path.realpath(__file__)),'html/unsupported_platform.html'))
-
-    async def handle_login_failed_get(self, request):
-        return aiohttp.web.FileResponse(os.path.join(os.path.dirname(os.path.realpath(__file__)),'html/login_failed.html'))
-
-    async def handle_2fa_get(self, request):
-        return aiohttp.web.FileResponse(os.path.join(os.path.dirname(os.path.realpath(__file__)),'html/2fa.html'))
-
-    async def handle_2fa_failed_get(self, request):
-        return aiohttp.web.FileResponse(os.path.join(os.path.dirname(os.path.realpath(__file__)),'html/2fa_failed.html'))
-
-    async def handle_finished_get(self, request):
-        return aiohttp.web.FileResponse(os.path.join(os.path.dirname(os.path.realpath(__file__)),'html/finished.html'))
-
-    async def handle_login_post(self, request):
-        data = await request.post()
-        auth_result = WGCAuthorizationResult.FAILED
-
-        #check data
-        data_valid = True
-        if 'realm' not in data or not data['realm']:
-            logging.warning('wgc_auth_server/handle_login_post: data is not valid, realm is missing')
-            data_valid = False
-        if 'email' not in data or not data['email']:
-            logging.warning('wgc_auth_server/handle_login_post: data is not valid, email is missing')
-            data_valid = False
-        if 'password' not in data or not data['password']:
-            logging.warning('wgc_auth_server/handle_login_post: data is not valid, password is missing')
-            data_valid = False
-
-        if data_valid:
-            auth_result = await self.__backend.do_auth_emailpass(data['realm'], data['email'], data['password'])
-
-        self.__process_auth_result(auth_result)
-
-
-    async def handle_2fa_post(self, request):
-        data = await request.post()
-
-        auth_result = WGCAuthorizationResult.INCORRECT_2FA
-        if 'authcode' in data and data['authcode']:
-            use_backup_code = True if 'use_backup' in data else False
-            auth_result = await self.__backend.do_auth_2fa(data['authcode'], use_backup_code)
-
-        self.__process_auth_result(auth_result)
-
-
-    def __process_auth_result(self, auth_result):
-        if auth_result == WGCAuthorizationResult.FINISHED:
-            raise aiohttp.web.HTTPFound('/finished')
-        elif auth_result == WGCAuthorizationResult.REQUIRES_2FA:
-            raise aiohttp.web.HTTPFound('/2fa')
-        elif auth_result == WGCAuthorizationResult.INCORRECT_2FA or auth_result == WGCAuthorizationResult.INCORRECT_2FA_BACKUP: 
-            raise aiohttp.web.HTTPFound('/2fa_failed')
-        else:
-            raise aiohttp.web.HTTPFound('/login_failed')
+from .wgc_http import WgcHttp
 
 class WGCApi:
-    HTTP_USER_AGENT = 'wgc/20.01.00.9514'
-
     OAUTH_GRANT_TYPE_BYPASSWORD = 'urn:wargaming:params:oauth:grant-type:basic'
     OAUTH_GRANT_TYPE_BYTOKEN = 'urn:wargaming:params:oauth:grant-type:access-token'
     OUATH_URL_CHALLENGE = '/id/api/v2/account/credentials/create/oauth/token/challenge/'
@@ -136,7 +40,11 @@ class WGCApi:
     LOCALSERVER_HOST = '127.0.0.1'
     LOCALSERVER_PORT = 13337
 
-    def __init__(self, tracking_id : str = '', country_code : str = '', language_code : str = 'en'):
+    def __init__(self, http : WgcHttp, tracking_id : str = '', country_code : str = '', language_code : str = 'en'):
+        self.__logger = logging.getLogger('wgc_api')
+
+        self.__http = http
+
         self._tracking_id = tracking_id
         self._country_code = country_code
         self._language_code = language_code
@@ -146,47 +54,10 @@ class WGCApi:
         self.__login_info = None
         self.__login_info_temp = None
 
-        #aiohhtp
-        self.__sslcontext = ssl.create_default_context(cafile=certifi.where())
-        self.__connector = aiohttp.TCPConnector(ssl_context=self.__sslcontext)
-        self.__session_headers = {'User-Agent': self.HTTP_USER_AGENT}
-        self.__session = aiohttp.ClientSession(connector=self.__connector, headers = self.__session_headers)
 
     async def shutdown(self):
         if self.__server is not None:
             await self.auth_server_stop()
-        
-        await self.__session.close()
-
-    #
-    # Requests
-    #
-
-    async def request(self, method: str, url: str, *, params: Any = None, data: Any = None, json: Any = None):
-        response_status = 202
-        response_text = None
-
-        if 'Referer' in self.__session_headers:
-            self.__session_headers.pop('Referer')
-    
-        while True:
-            async with self.__session.request(method, url, headers = self.__session_headers, params = params, data = data, json = json) as response:
-                response_text = await response.text()
-                response_status = response.status
-                if response_status == 202 and 'Location' in response.headers:
-                    url = response.headers['Location']
-                    self.__session_headers.update({'Referer': str(response.url)})
-                    method = 'GET'
-                else:
-                    break
-
-        return namedtuple('WgcResponse', ['status', 'text'])(response_status, response_text)
-
-    async def request_get(self, url: str) -> Any:
-        return await self.request('GET', url)
-
-    async def request_post(self, url: str, *, params: Any = None, data: Any = None, json: Any = None) -> Any:
-        return await self.request('POST', url, params = params, data = data, json = json)
 
     # 
     # Getters
@@ -194,44 +65,44 @@ class WGCApi:
 
     def get_account_id(self) -> int:
         if self.__login_info is None:
-            logging.error('login info is none')
+            self.__logger.error('get_account_id: login info is none')
             return None
 
         if 'user' not in self.__login_info:
-            logging.error('login info does not contains user id')
+            self.__logger.error('get_account_id: login info does not contains user id')
             return None
 
         return int(self.__login_info['user'])
 
     def get_account_email(self) -> str:
         if self.__login_info is None:
-            logging.error('login info is none')
+            self.__logger.error('get_account_email: login info is none')
             return None
 
         if 'email' not in self.__login_info:
-            logging.error('login info does not contains email')
+            self.__logger.error('get_account_email: login info does not contains email')
             return None
 
         return self.__login_info['email']
 
     def get_account_nickname(self) -> str:
         if self.__login_info is None:
-            logging.error('login info is none')
+            self.__logger.error('get_account_nickname: login info is none')
             return None
 
         if 'nickname' not in self.__login_info:
-            logging.error('login info does not contains nickname')
+            self.__logger.error('get_account_nickname: login info does not contains nickname')
             return None
 
         return self.__login_info['nickname']
         
     def get_account_realm(self) -> str:
         if self.__login_info is None:
-            logging.error('login info is none')
+            self.__logger.error('get_account_realm: login info is none')
             return None
 
         if 'realm' not in self.__login_info:
-            logging.error('login info does not contains realm')
+            self.__logger.error('get_account_realm: login info does not contains realm')
             return None
 
         return self.__login_info['realm']
@@ -451,7 +322,7 @@ class WGCApi:
             logging.error('wgc_auth/update_bearer: login info does not contain exchange code')
             return None
 
-        self.__session_headers.update({'Authorization':'Bearer %s:%s' % (self.__login_info['access_token'], self.__login_info['exchange_code'])})
+        self.__http.update_headers({'Authorization':'Bearer %s:%s' % (self.__login_info['access_token'], self.__login_info['exchange_code'])})
 
     #
     # URL formatting
@@ -480,7 +351,7 @@ class WGCApi:
             return None
 
     async def __oauth_challenge_get(self, realm):
-        r = await self.request_get(self.__get_url('wgnet', realm, self.OUATH_URL_CHALLENGE))
+        r = await self.__http.request_get(self.__get_url('wgnet', realm, self.OUATH_URL_CHALLENGE))
         if r.status != 200:
             logging.error('wgc_auth/oauth_challenge_get: error %s, content: %s' % (r.status, r.text))
             return None
@@ -528,7 +399,7 @@ class WGCApi:
             else:
                 body['otp_code'] = otp_code
 
-        response = await self.request_post(self.__get_url('wgnet', realm, self.OAUTH_URL_TOKEN), data = body)
+        response = await self.__http.request_post(self.__get_url('wgnet', realm, self.OAUTH_URL_TOKEN), data = body)
         
         result = None
         try:
@@ -548,7 +419,7 @@ class WGCApi:
         body['exchange_code'] = ''.join(random.choices(string.digits+'ABCDEF', k=32))
         body['tid'] = self._tracking_id
 
-        response = await self.request_post(self.__get_url('wgnet', realm, self.OAUTH_URL_TOKEN), data = body)
+        response = await self.__http.request_post(self.__get_url('wgnet', realm, self.OAUTH_URL_TOKEN), data = body)
 
         if response.status != 200:
             logging.error('wgc_auth/__oauth_token_get_bytoken: error on receiving token by token: %s because status is %s' % (response.text, response.status))
@@ -589,7 +460,7 @@ class WGCApi:
             logging.error('wgc_api/__wgni_create_token_1: login info does not contain access_token')
             return None
 
-        response = await self.request_post(
+        response = await self.__http.request_post(
             self.__get_url('wgnet', self.__login_info['realm'], self.WGNI_URL_TOKEN1), 
             data = { 'requested_for' : requested_for, 'access_token' : self.__login_info['access_token'] })
 
@@ -612,7 +483,7 @@ class WGCApi:
             logging.error('wgc_auth/wgni_get_account_info: login info does not contain realm')
             return None
 
-        response = await self.request_post(
+        response = await self.__http.request_post(
             self.__get_url('wgnet', self.__login_info['realm'], self.WGNI_URL_ACCOUNTINFO), 
             data = { 'fields' : 'nickname' })
         
@@ -674,7 +545,7 @@ class WGCApi:
             logging.error('wgc_auth/__wgcps_fetch_product_list: login info does not contain realm')
             return None
 
-        response = await self.request_post(
+        response = await self.__http.request_post(
             self.__get_url('wgcps', self.__login_info['realm'], self.WGCPS_FETCH_PRODUCT_INFO), 
             json = { 'account_id' : self.get_account_id(), 'country' : self._country_code, 'storefront' : 'wgc_showcase' })
 
@@ -704,7 +575,7 @@ class WGCApi:
         #load additional adata
         response_content['data']['product_content'] = list()
         for product_uri in response_content['data']['product_uris']:
-            product_response = await self.request_get(product_uri)
+            product_response = await self.__http.request_get(product_uri)
             if product_response.status != 200:
                 logging.error('wgc_auth/__wgcps_fetch_product_list: error on retrieving product info: %s' % product_uri)
                 continue
@@ -732,7 +603,7 @@ class WGCApi:
         url = url + '&country_code=%s' % self._country_code
         url = url + additionals
 
-        showroom_response = await self.request_get(url)
+        showroom_response = await self.__http.request_get(url)
         
         if showroom_response.status != 200:
             logging.error('wgc_auth/__wguscs_get_showroom: error on retrieving showroom data: %s' % showroom_response.text)
